lucet-wiggle: reorganize crate to be like upstream wiggle.
the `lucet-wiggle` crate is now a big umbrella, mimicking the
structure of wiggle:
* defines the runtime components
* re-exports the proc macro defined in `lucet-wiggle/macro`
* re-exports `wiggle` itself. This way, consumer crates of lucet_wasi
  don't have to depend on wiggle (or witx) explicitly. Less version
  bounds to manage.
* re-exports the lucet-wiggle/generator (mostly useful for bindings,
  also for lucet-wasi-generate)
