+++ lucet-wasi/Cargo.toml
@@ -18,9 +18,7 @@ lucet-runtime = { path = "../lucet-runtime", version = "=0.6.2-dev"  }
 lucet-runtime-internals = { path = "../lucet-runtime/lucet-runtime-internals", version = "=0.6.2-dev" }
 lucet-module = { path = "../lucet-module", version = "=0.6.2-dev" }
 lucet-wasi-generate = { path = "./generate", version = "=0.6.2-dev" }
-lucet-wiggle-generate = { path = "../lucet-wiggle/generate", version = "=0.6.2-dev"  }
-lucet-wiggle-runtime = { path = "../lucet-wiggle/runtime", version = "=0.6.2-dev" }
-wiggle = { path = "../wasmtime/crates/wiggle", version = "0.15.0" }
+lucet-wiggle = { path = "../lucet-wiggle", version = "=0.6.2-dev" }
 libc = "0.2.65"
 nix = "0.17"
 rand = "0.6"
