--- lucet-wiggle/runtime/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "lucet-wiggle-runtime"
-version = "0.6.2-dev"
-homepage = "https://github.com/fastly/lucet"
-repository = "https://github.com/fastly/lucet"
-license = "Apache-2.0 WITH LLVM-exception"
-categories = ["wasm"]
-authors = ["Lucet team <lucet@fastly.com>"]
-edition = "2018"
-
-[dependencies]
-lucet-runtime = { path = "../../lucet-runtime", version = "0.6.2-dev" }
-wiggle =  { path = "../../wasmtime/crates/wiggle", version = "0.15.0" }
